Warning fixes.
XFS replicates quite a bit of X server infrastructure so that it can
share libXfont. Many of those shared functions are declared in shared
font header files (either fontsproto.h or fontmisc.h). This patch
removes duplicate definitions from the XFS header files and makes sure
the shared header files are included in the right places to get the
functions declared before being defined.

Of course, some of those function types have changed, in particular
CopyISOLatin1Lowered now takes a const char * source parameter.

This patch also stops re-using the 'port' variable in
GetInetdListenInfo when fetching the ReopenInfo. That's not strictly
necessary here, but will be useful when the Reopen API in libxtrans
changes to take a constant parameter.

Reviewed-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Tested-by: Gaetan Nadon <memsize@videotron.ca>
Signed-off-by: Keith Packard <keithp@keithp.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>